requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
colorama==0.4.6
cryptography==46.0.6
exceptiongroup==1.3.1
h11==0.16.0
idna==3.11
iniconfig==2.3.0
outcome==1.3.0.post0
packaging==26.0
pluggy==1.6.0
pycparser==3.0
Pygments==2.20.0
pyOpenSSL==26.0.0
PySocks==1.7.1
pytest==9.0.2
selenium==4.41.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.4.1
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.3
urllib3-secure-extra==0.1.0
websocket-client==1.9.0
wsproto==1.3.2
